Wrap async validation call
For some reason, inlining the call to `isExistingContributingFactor` was appearing to return false even in cases where it should return true. It might just be something weird with the Remix dev server (HMR stuff), but this seems to fix it so whatever.

Also removed the optional chaining, which was mistakenly leftover from previous attempts at async validation.